pyproject.toml
[tool.poetry]
name = "{PoketDevTools}"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^6.0.0"
black = "^22.12.0"
radon = "^5.1.0"
bandit = "^1.7.4"
isort = "^5.11.4"
mypy = "^0.991"
pycairo = "^1.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
